distributed-systems/assignment-1/grpc-clientserver/requirements.txt
grpcio==1.47.0
grpcio-tools==1.47.0
protobuf==3.20.1
sqlmodel==0.0.11
pytest==7.4.0